+++ app/src/main/java/app/kaito_dogi/mybrary/MybraryNavHost.kt
@@ -7,9 +7,10 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import app.kaito_dogi.mybrary.core.ui.navigation.MybraryRoute
 import app.kaito_dogi.mybrary.feature.searchbook.searchBookScreen
-import app.kaito_dogi.mybrary.feature.sendotp.sendOtpScreen
-import app.kaito_dogi.mybrary.feature.verifyotp.verifyOtpScreen
+import app.kaito_dogi.mybrary.feature.auth.destination.sendotp.sendOtpScreen
+import app.kaito_dogi.mybrary.feature.auth.destination.verifyotp.verifyOtpScreen
 
+@Deprecated("")
 @Composable
 internal fun MybraryNavHost(
   modifier: Modifier = Modifier,
@@ -18,44 +19,9 @@ internal fun MybraryNavHost(
 
   NavHost(
     navController = navController,
-    startDestination = MybraryRoute.SendOtp,
+    startDestination = MybraryRoute.SearchBook,
     modifier = modifier.fillMaxSize(),
   ) {
     searchBookScreen()
-
-    sendOtpScreen(
-      onSendOtpComplete = { email, page ->
-        val route = MybraryRoute.VerifyOtp(
-          email = email,
-          page = page,
-        )
-        navController.navigate(route)
-      },
-      onLoginComplete = {
-        navController.navigate(MybraryRoute.MyBookList) {
-          popUpTo<MybraryRoute.SendOtp> {
-            inclusive = true
-          }
-        }
-      },
-      onSignUpComplete = {
-        navController.navigate(MybraryRoute.MyBookList) {
-          popUpTo<MybraryRoute.SendOtp> {
-            inclusive = true
-          }
-        }
-      },
-    )
-
-    verifyOtpScreen(
-      onVerifyOtpComplete = {
-        navController.navigate(MybraryRoute.MyBookList) {
-          popUpTo<MybraryRoute.SendOtp> {
-            inclusive = true
-          }
-        }
-      },
-      onNavigationIconClick = navController::popBackStack,
-    )
   }
 }
