+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/navigation/bar/MainNavigation.kt
@@ -14,4 +14,4 @@ fun NavGraphBuilder.mainNavGraph(
   builder = builder,
 )
 
-fun NavHostController.navigateToMainNavGraph() = this.navigate(AppRoute.Main)
+fun NavHostController.navigateToMainNavGraph() = this.navigate(route = AppRoute.Main)
