+++ feature/auth/src/main/java/app/kaito_dogi/mybrary/feature/auth/AuthNavigation.kt
@@ -1,0 +1,14 @@
+package app.kaito_dogi.mybrary.feature.auth
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.navigation
+import app.kaito_dogi.mybrary.core.ui.navigation.route.AppRoute
+import app.kaito_dogi.mybrary.core.ui.navigation.route.AuthRoute
+
+fun NavGraphBuilder.authNavGraph(
+  startDestination: AuthRoute,
+  builder: NavGraphBuilder.() -> Unit,
+) = navigation<AppRoute.Auth>(
+  startDestination = startDestination,
+  builder = builder,
+)
