+++ feature/my-book/src/main/java/app/kaito_dogi/mybrary/feature/mybook/destination/mybooklist/MyBookListNavigation.kt
@@ -4,6 +4,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import app.kaito_dogi.mybrary.core.domain.model.MyBook
+import app.kaito_dogi.mybrary.core.ui.navigation.route.AuthRoute
 import app.kaito_dogi.mybrary.core.ui.navigation.route.MyBookRoute
 
 fun NavGraphBuilder.myBookListScreen(
@@ -16,4 +17,8 @@ fun NavGraphBuilder.myBookListScreen(
   )
 }
 
-fun NavHostController.navigateToMyBookListScreen() = this.navigate(route = MyBookRoute.MyBookList)
+fun NavHostController.navigateToMyBookListScreen() = this.navigate(route = MyBookRoute.MyBookList) {
+  popUpTo<AuthRoute.SendOtp> {
+    inclusive = true
+  }
+}
