+++ feature/setting/src/main/java/app/kaito_dogi/mybrary/feature/setting/destination/settinglist/SettingListNavigation.kt
@@ -10,4 +10,4 @@ fun NavGraphBuilder.settingListScreen(
   SettingListScreenContainer()
 }
 
-fun NavHostController.navigateToSettingListScreen() = this.navigate(SettingRoute.SettingList)
+fun NavHostController.navigateToSettingListScreen() = this.navigate(route = SettingRoute.SettingList)
