+++ settings.gradle.kts
@@ -25,8 +25,7 @@ include(":core:design-system")
 include(":core:domain")
 include(":core:supabase")
 include(":core:ui")
+include(":feature:auth")
 include(":feature:my-book")
 include(":feature:search-book")
-include(":feature:send-otp")
 include(":feature:setting")
-include(":feature:verify-otp")
